// src/app/api/model/menuResponseDTO.ts
/**
 * Inventory AUTH API - JWT Secured
 *
 * Contact: adrian2x2001@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MenuResponseDTO { 
    id?: number;
    label?: string;
    description?: string;
    route?: string;
    parentId?: number;
    active?: boolean;
    icon?: string;
    order?: number;
    creationDate?: string;
    lastModificationDate?: string;
    internalRoute?: boolean;
    applicationId?: number;
    children?: Array<any>;
}