// src/app/api/model/roleRequestDTO.ts
/**
 * Inventory AUTH API - JWT Secured
 *
 * Contact: adrian2x2001@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface RoleRequestDTO { 
    name: string;
    description?: string;
    applicationId: number;
    privilegesId?: Set<number>;
    menusId?: Set<number>;
}